requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
scipy>=1.10.0

# Image processing
opencv-python>=4.8.0
Pillow>=10.0.0
imageio>=2.31.0
imageio-ffmpeg>=0.4.9

# Visualization
matplotlib>=3.7.0
plotly>=5.14.0

# 3D processing
open3d>=0.17.0
trimesh>=3.21.0
plyfile>=1.0.0

# Utilities
tqdm>=4.65.0
pyyaml>=6.0
omegaconf>=2.3.0
einops>=0.6.1
hydra-core>=1.3.0

# Deep learning utilities
timm>=0.9.0
transformers>=4.30.0
diffusers>=0.21.0
accelerate>=0.20.0

# 3D Gaussian Splatting
# Note: Install from source after basic deps
# git+https://github.com/graphdeco-inria/gaussian-splatting.git
# git+https://github.com/graphdeco-inria/diff-gaussian-rasterization.git

# Scientific computing
scikit-image>=0.21.0
scikit-learn>=1.3.0

# Logging and experiment tracking
tensorboard>=2.13.0
wandb>=0.15.0
